Use set_bits from transform::util
