+++ arrow/Cargo.toml
@@ -105,6 +105,10 @@ name = "boolean_kernels"
 harness = false
 
 [[bench]]
+name = "boolean_append_packed"
+harness = false
+
+[[bench]]
 name = "arithmetic_kernels"
 harness = false
