+++ arrow/src/util/mod.rs
@@ -18,6 +18,7 @@
 #[cfg(feature = "test_utils")]
 pub mod bench_util;
 pub mod bit_chunk_iterator;
+pub(crate) mod bit_mask;
 pub mod bit_util;
 #[cfg(feature = "test_utils")]
 pub mod data_gen;
